# backend/docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "chi (net/http) + pgx (PostgreSQL) + Redis дээр суурилсан Clean Architecture бүхий Go backend. Нээлттэй эхийн snykk/go-rest-boilerplate (MIT, зохиогч Najib Fikri)-ээс үүсэлтэй; HTTP давхаргыг chi, өгөгдлийн давхаргыг pgx руу хөрвүүлсэн.",
        "title": "Government Template Platform V3.0 API",
        "termsOfService": "https://github.com/snykk/go-rest-boilerplate",
        "contact": {
            "name": "Government Template Platform V3.0",
            "url": "https://github.com/snykk/go-rest-boilerplate"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/gov/applications": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "gov"
                ],
                "summary": "Миний хүсэлтүүд",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/template_internal_http_handlers_v1.BaseResponse"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "gov"
                ],
                "summary": "Үйлчилгээнд хүсэлт гаргах",
                "parameters": [
                    {
                        "description": "Apply",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/template_internal_http_datatransfers_requests.GovApplyRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/template_internal_http_handlers_v1.BaseResponse"
                        }
                    }
                }
            }
        },
        "/gov/applications/{id}/cancel": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "gov"
                ],
                "summary": "Хүсэлт цуцлах",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Application ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/template_internal_http_handlers_v1.BaseResponse"
                        }
                    }
                }
            }
        },
        "/gov/appointments": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "gov"
                ],
                "summary": "Цаг захиалгууд",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/template_internal_http_handlers_v1.BaseResponse"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "gov"
                ],
                "summary": "Цаг захиалах",
                "parameters": [
                    {
                        "description": "Book",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/template_internal_http_datatransfers_requests.GovBookRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/template_internal_http_handlers_v1.BaseResponse"
                        }
                    }
                }
            }
        },
        "/gov/appointments/{id}/cancel": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "gov"
                ],
                "summary": "Цаг захиалга цуцлах",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Appointment ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/template_internal_http_handlers_v1.BaseResponse"
                        }
                    }
                }
            }
        },
...
        "/gov/notifications/{id}/read": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "gov"
                ],
                "summary": "Мэдэгдлийг уншсан болгох",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Notification ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/template_internal_http_handlers_v1.BaseResponse"
                        }
                    }
                }
            }
        },
...
        "/gov/payments/{id}/pay": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "gov"
                ],
                "summary": "Төлбөр төлөх",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Payment ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/template_internal_http_handlers_v1.BaseResponse"
                        }
                    }
                }
            }
        },
        "/gov/references": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "gov"
                ],
                "summary": "Миний лавлагаа",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/template_internal_http_handlers_v1.BaseResponse"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "gov"
                ],
                "summary": "Лавлагаа захиалах",
                "parameters": [
                    {
                        "description": "Reference",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/template_internal_http_datatransfers_requests.GovReferenceRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/template_internal_http_handlers_v1.BaseResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "template_internal_http_datatransfers_requests.GovApplyRequest": {
            "type": "object",
            "required": [
                "service_id"
            ],
            "properties": {
                "note": {
                    "type": "string",
                    "maxLength": 500
                },
                "service_id": {
                    "type": "string"
                }
            }
        },
        "template_internal_http_datatransfers_requests.GovBookRequest": {
            "type": "object",
            "required": [
                "scheduled_at"
            ],
            "properties": {
                "location": {
                    "type": "string",
                    "maxLength": 200
                },
                "note": {
                    "type": "string",
                    "maxLength": 500
                },
                "scheduled_at": {
                    "type": "string"
                },
                "service_id": {
                    "type": "string"
                }
            }
        },
        "template_internal_http_datatransfers_requests.GovReferenceRequest": {
            "type": "object",
            "required": [
                "type"
            ],
            "properties": {
                "type": {
                    "type": "string",
                    "enum": [
                        "residence",
                        "birth",
                        "marriage",
                        "tax",
                        "social_ins",
                        "criminal"
                    ]
                }
            }
        },
...
        "template_internal_http_handlers_v1.BaseResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                },
                "request_id": {
                    "type": "string"
                },
                "status": {
                    "type": "boolean"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "/auth/login эсвэл /auth/refresh-аас олгогдсон Bearer хандах токен (access token).",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}